feat: Redesigns Landing page. Improves Components structure
