--- src/components/atoms/index.ts
@@ -1,4 +1,0 @@
-export * from "./app-button";
-export * from "./app-checkbox";
-export * from "./app-timer-bar";
-export * from "./app-timer-countdown";
