--- src/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./atoms";
-export * from "./molecules";
-export * from "./organisms";
-export * from "./smart";
