--- src/components/molecules/game-question/index.ts
@@ -1,1 +1,0 @@
-export * from "./game-question";
