--- src/components/organisms/app-menu-overlay/app-menu-overlay.tsx
@@ -1,34 +1,0 @@
-import { XMarkIcon } from "@heroicons/react/24/outline";
-import { AppButton } from "../../atoms";
-import { cn } from "@/utils";
-import { ReactNode } from "react";
-
-interface AppMenuOverlayProps {
-  children?: ReactNode;
-  isOpen?: boolean;
-  onClose: () => void;
-}
-
-export const AppMenuOverlay = (props: AppMenuOverlayProps) => {
-  const { isOpen, onClose, children } = props;
-
-  return (
-    <nav
-      className={cn(
-        "fixed h-screen z-50 inset-0 bg-black duration-300",
-        !isOpen && "opacity-0 pointer-events-none"
-      )}
-    >
-      <div className="max-w-screen-2xl h-full mx-auto flex flex-col">
-        <div className="flex justify-end py-4 px-8">
-          <AppButton
-            variant="icon"
-            prefixIcon={<XMarkIcon className="size-8" />}
-            onClick={onClose}
-          />
-        </div>
-        {children}
-      </div>
-    </nav>
-  );
-};
