--- src/components/organisms/index.ts
@@ -1,2 +1,0 @@
-export * from "./app-menu-overlay";
-export * from "./app-navbar";
