+++ src/components/ui/app-footer/app-footer.tsx
@@ -1,0 +1,10 @@
+export const AppFooter = () => {
+  return (
+    <footer className="flex justify-center items-center gap-2 py-4">
+      <p className="font-light text-sm space-x-2">Developed by BurmesePotato</p>
+      <span role="img" aria-label="potato">
+        🥔
+      </span>
+    </footer>
+  );
+};
