+++ src/components/ui/app-question/app-question.tsx
@@ -1,0 +1,15 @@
+interface AppQuestionProps {
+  question: string;
+  letter: string;
+}
+
+export const AppQuestion = (props: AppQuestionProps) => {
+  const { question, letter } = props;
+
+  return (
+    <section className="flex flex-col gap-8">
+      <p className="text-2xl">{question}</p>
+      <p className="text-7xl font-bold text-center uppercase">{letter}</p>
+    </section>
+  );
+};
