+++ src/components/ui/app-question/index.ts
@@ -1,0 +1,1 @@
+export * from "./app-question";
