+++ src/components/widgets/game-cell/index.ts
