+++ src/components/widgets/index.ts
