+++ src/features/home/index.ts
@@ -1,0 +1,3 @@
+export * from "./section-hero";
+export * from "./section-learnModes";
+export * from "./section-roadmap";
