--- src/features/landing/index.ts
@@ -1,2 +1,0 @@
-export * from "./landing-quiz";
-export * from "./landing-practice";
