--- src/features/landing/landing-practice/index.ts
@@ -1,1 +1,0 @@
-export * from "./landing-practice";
