--- src/features/landing/landing-quiz/index.ts
@@ -1,1 +1,0 @@
-export * from "./landing-quiz";
