+++ src/pages/quiz/quiz-choose-page.tsx
@@ -1,4 +1,4 @@
-import { AppButton, AppNavbar } from "@/components";
+import { AppButton, AppNavbar } from "@/components/ui";
 import { useNavigate } from "react-router-dom";
 import { GameMode } from "@/features/shared/types";
 
@@ -23,9 +23,9 @@ export default function QuizChoosePage() {
   };
 
   return (
-    <main className="max-w-screen-2xl mx-auto px-8">
-      <div className="flex flex-col min-h-screen">
-        <AppNavbar backBtnText="Back to Main Menu" onBack={handleBack} />
+    <main className="max-w-screen-xl mx-auto">
+      <div className="flex flex-col min-h-screen px-8">
+        <AppNavbar backBtnText="Main Menu" onBack={handleBack} />
         <section className="flex flex-col items-center gap-16 grow justify-center">
           <h1 className="text-4xl font-bold">Choose Game Mode</h1>
           <div className="flex flex-col gap-4">
@@ -34,11 +34,9 @@ export default function QuizChoosePage() {
               onClick={() => handleSelect("LIFE_SMASH")}
               size="lg"
             /> */}
-            <AppButton
-              label="Timer Smash"
-              onClick={() => handleSelect("TIMER_SMASH")}
-              size="lg"
-            />
+            <AppButton onClick={() => handleSelect("TIMER_SMASH")} size="lg">
+              Timer Smash
+            </AppButton>
           </div>
         </section>
       </div>
