+++ src/styles/global.css
@@ -14,15 +14,12 @@
 }
 
 body {
-  background-color: #0d1013;
-  color: #fff;
-  margin: 0;
-  min-height: 100vh;
+  @apply min-h-screen m-0 bg-slate-900 text-white;
 }
 
 @media (prefers-color-scheme: light) {
 }
 
 .bento {
-  @apply rounded-[3rem] p-10 border border-2 border-white;
+  @apply rounded-3xl p-8 border border-white;
 }
