+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/components/index.ts","./src/components/atoms/index.ts","./src/components/atoms/app-button/app-button.tsx","./src/components/atoms/app-button/index.ts","./src/components/atoms/app-checkbox/app-checkbox.tsx","./src/components/atoms/app-checkbox/index.ts","./src/components/atoms/app-timer-bar/app-timer-bar.tsx","./src/components/atoms/app-timer-bar/index.ts","./src/components/atoms/app-timer-countdown/app-timer-countdown.tsx","./src/components/atoms/app-timer-countdown/index.ts","./src/components/molecules/index.ts","./src/components/molecules/game-question/game-question.tsx","./src/components/molecules/game-question/index.ts","./src/components/organisms/index.ts","./src/components/organisms/app-menu-overlay/app-menu-overlay.tsx","./src/components/organisms/app-menu-overlay/index.ts","./src/components/organisms/app-navbar/app-navbar.tsx","./src/components/organisms/app-navbar/index.ts","./src/components/smart/index.ts","./src/components/smart/game-cell/game-cell.tsx","./src/components/smart/game-cell/index.ts","./src/components/smart/game-menu/game-menu.tsx","./src/components/smart/game-menu/index.ts","./src/components/smart/game-sound-button/game-sound-button.tsx","./src/components/smart/game-sound-button/index.ts","./src/components/smart/game-timer/game-timer.tsx","./src/components/smart/game-timer/index.ts","./src/features/landing/index.ts","./src/features/landing/landing-practice/index.ts","./src/features/landing/landing-practice/landing-practice.tsx","./src/features/landing/landing-quiz/index.ts","./src/features/landing/landing-quiz/landing-quiz.tsx","./src/features/quiz/index.ts","./src/features/shared/constants/characters.ts","./src/features/shared/constants/game-config.ts","./src/features/shared/constants/index.ts","./src/features/shared/hooks/usetimersmash.ts","./src/features/shared/states/index.ts","./src/features/shared/states/quiz-game.ts","./src/features/shared/states/timer-smash.ts","./src/features/shared/types/braille.ts","./src/features/shared/types/game.ts","./src/features/shared/types/index.ts","./src/features/shared/ui/smart/game-menu/game-menu.tsx","./src/hooks/index.ts","./src/hooks/usecharacter.ts","./src/hooks/usetimersmash.ts","./src/layouts/index.ts","./src/layouts/quiz-layout.tsx","./src/libs/router.ts","./src/pages/index.ts","./src/pages/landing/index.tsx","./src/pages/landing/landing-page.tsx","./src/pages/practice/index.tsx","./src/pages/practice/practice-page.tsx","./src/pages/quiz/index.tsx","./src/pages/quiz/quiz-choose-page.tsx","./src/pages/quiz/timer-smash-page.tsx","./src/routes/root-routes.tsx","./src/utils/index.ts","./src/utils/utils.ts"],"version":"5.6.2"}
+{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/components/ui/index.ts","./src/components/ui/app-button/app-button.tsx","./src/components/ui/app-button/index.ts","./src/components/ui/app-checkbox/app-checkbox.tsx","./src/components/ui/app-checkbox/index.ts","./src/components/ui/app-footer/app-footer.tsx","./src/components/ui/app-footer/index.ts","./src/components/ui/app-menu-overlay/app-menu-overlay.tsx","./src/components/ui/app-menu-overlay/index.ts","./src/components/ui/app-navbar/app-navbar.tsx","./src/components/ui/app-navbar/index.ts","./src/components/ui/app-question/app-question.tsx","./src/components/ui/app-question/index.ts","./src/components/ui/app-timer-bar/app-timer-bar.tsx","./src/components/ui/app-timer-bar/index.ts","./src/components/ui/app-timer-countdown/app-timer-countdown.tsx","./src/components/ui/app-timer-countdown/index.ts","./src/components/widgets/index.ts","./src/components/widgets/game-cell/game-cell.tsx","./src/components/widgets/game-cell/index.ts","./src/components/widgets/game-menu/game-menu.tsx","./src/components/widgets/game-menu/index.ts","./src/components/widgets/game-sound-button/game-sound-button.tsx","./src/components/widgets/game-sound-button/index.ts","./src/components/widgets/game-timer/game-timer.tsx","./src/components/widgets/game-timer/index.ts","./src/features/home/index.ts","./src/features/home/section-hero.tsx","./src/features/home/section-learnmodes.tsx","./src/features/home/section-roadmap.tsx","./src/features/quiz/index.ts","./src/features/quiz/timer-smash/timer-smash.tsx","./src/features/shared/constants/characters.ts","./src/features/shared/constants/game-config.ts","./src/features/shared/constants/index.ts","./src/features/shared/hooks/usetimersmash.ts","./src/features/shared/states/index.ts","./src/features/shared/states/quiz-game.ts","./src/features/shared/states/timer-smash.ts","./src/features/shared/types/braille.ts","./src/features/shared/types/game.ts","./src/features/shared/types/index.ts","./src/features/shared/ui/smart/game-menu/game-menu.tsx","./src/hooks/index.ts","./src/hooks/usecharacter.ts","./src/hooks/usetimersmash.ts","./src/layouts/index.ts","./src/layouts/quiz-layout.tsx","./src/libs/router.ts","./src/pages/index.ts","./src/pages/landing/index.tsx","./src/pages/landing/landing-page.tsx","./src/pages/practice/index.tsx","./src/pages/practice/practice-page.tsx","./src/pages/quiz/index.tsx","./src/pages/quiz/quiz-choose-page.tsx","./src/pages/quiz/quiz-play-page.tsx","./src/pages/quiz/timer-smash-page.tsx","./src/routes/root-routes.tsx","./src/utils/index.ts","./src/utils/utils.ts"],"version":"5.6.2"}
